Move shared structs to mod.rs
